remove preview env refs for now
