--- api/server/router/preview_environment.go
@@ -1,86 +1,0 @@
-package router
-
-import (
-	"github.com/go-chi/chi/v5"
-	"github.com/porter-dev/porter/api/server/handlers/preview_environment"
-	"github.com/porter-dev/porter/api/server/shared"
-	"github.com/porter-dev/porter/api/server/shared/config"
-	"github.com/porter-dev/porter/api/server/shared/router"
-	"github.com/porter-dev/porter/api/types"
-)
-
-func NewPreviewEnvironmentScopedRegisterer(children ...*router.Registerer) *router.Registerer {
-	return &router.Registerer{
-		GetRoutes: GetPreviewEnvironmentScopedRoutes,
-		Children:  children,
-	}
-}
-
-func GetPreviewEnvironmentScopedRoutes(
-	r chi.Router,
-	config *config.Config,
-	basePath *types.Path,
-	factory shared.APIEndpointFactory,
-	children ...*router.Registerer,
-) []*router.Route {
-	routes, projPath := getPreviewEnvRoutes(r, config, basePath, factory)
-
-	if len(children) > 0 {
-		r.Route(projPath.RelativePath, func(r chi.Router) {
-			for _, child := range children {
-				childRoutes := child.GetRoutes(r, config, basePath, factory, child.Children...)
-
-				routes = append(routes, childRoutes...)
-			}
-		})
-	}
-
-	return routes
-}
-
-func getPreviewEnvRoutes(
-	r chi.Router,
-	config *config.Config,
-	basePath *types.Path,
-	factory shared.APIEndpointFactory,
-) ([]*router.Route, *types.Path) {
-	relPath := "/preview_environments"
-
-	newPath := &types.Path{
-		Parent:       basePath,
-		RelativePath: relPath,
-	}
-
-	var routes []*router.Route
-
-	// POST /api/projects/{project_id}/clusters/{cluster_id}/preview_environments -> preview_environment.NewCreatePreviewEnvironmentHandler
-	createPreviewEnvEndpoint := factory.NewAPIEndpoint(
-		&types.APIRequestMetadata{
-			Verb:   types.APIVerbCreate,
-			Method: types.HTTPVerbPost,
-			Path: &types.Path{
-				Parent:       basePath,
-				RelativePath: relPath,
-			},
-			Scopes: []types.PermissionScope{
-				types.UserScope,
-				types.ProjectScope,
-				types.ClusterScope,
-			},
-		},
-	)
-
-	createPreviewEnvHandler := preview_environment.NewCreatePreviewEnvironmentHandler(
-		config,
-		factory.GetDecoderValidator(),
-		factory.GetResultWriter(),
-	)
-
-	routes = append(routes, &router.Route{
-		Endpoint: createPreviewEnvEndpoint,
-		Handler:  createPreviewEnvHandler,
-		Router:   r,
-	})
-
-	return routes, newPath
-}
