+++ api/server/router/router.go
@@ -33,8 +33,7 @@ func NewAPIRouter(config *config.Config) *chi.Mux {
 	clusterIntegrationRegisterer := NewClusterIntegrationScopedRegisterer()
 	stackRegisterer := NewStackScopedRegisterer()
 	envConfigRegisterer := NewEnvConfigScopedRegisterer()
-	previewEnvRegisterer := NewPreviewEnvironmentScopedRegisterer()
-	clusterRegisterer := NewClusterScopedRegisterer(namespaceRegisterer, clusterIntegrationRegisterer, stackRegisterer, envConfigRegisterer, previewEnvRegisterer)
+	clusterRegisterer := NewClusterScopedRegisterer(namespaceRegisterer, clusterIntegrationRegisterer, stackRegisterer, envConfigRegisterer)
 	infraRegisterer := NewInfraScopedRegisterer()
 	gitInstallationRegisterer := NewGitInstallationScopedRegisterer()
 	registryRegisterer := NewRegistryScopedRegisterer()
