+++ cli/cmd/stack/hooks.go
@@ -117,27 +117,6 @@ func (t *DeployAppHook) applyApp(shouldCreate bool, driverOutput map[string]inte
 		}
 		return fmt.Errorf("error updating app %s: %w", t.ApplicationName, err)
 	}
-
-	if t.EnvironmentMeta.GitHubMetadata.BranchFrom != "" {
-		color.New(color.FgGreen).Printf("Creating preview environment for app %s based on branch '%s'\n", t.ApplicationName, t.EnvironmentMeta.GitHubMetadata.BranchFrom)
-		// create preview env record
-		_, err = t.Client.CreatePreviewEnvironment(
-			context.Background(),
-			t.ProjectID,
-			t.ClusterID,
-			&types.CreatePreviewEnvironmentRequest{
-				EnvironmentConfigID: t.EnvironmentMeta.EnvironmentConfigID,
-				GitRepoOwner:        t.EnvironmentMeta.GitHubMetadata.RepoOwner,
-				GitRepoName:         t.EnvironmentMeta.GitHubMetadata.Repo,
-				Branch:              t.EnvironmentMeta.GitHubMetadata.BranchFrom,
-			},
-		)
-
-		if err != nil {
-			return fmt.Errorf("error creating preview environment: %w", err)
-		}
-	}
-
 	return nil
 }
