+++ internal/repository/gorm/migrate.go
@@ -63,7 +63,6 @@ func AutoMigrate(db *gorm.DB, debug bool) error {
 		&models.PorterApp{},
 		&models.PorterAppEvent{},
 		&models.EnvironmentConfig{},
-		&models.PreviewEnvironment{},
 		&ints.KubeIntegration{},
 		&ints.BasicIntegration{},
 		&ints.OIDCIntegration{},
